Fix race on startup between compute and compute desc
Xenon latest snapshot synchronizes a single factory at a time,
to throttle network I/O. However, for services that prevent start
from happening, because they wait on a child in another factory,
this causes problems.

Not sure if we should support child to child availability, but for now
remove the validate compute was doing on the description.

I think the check was bogus anyway. Resources should only validate self
not referential integrity across the system. The resource tasks should
do validation and grooming