Removal of factory services for resource services
This is part one. Part two will be removing factories
from task services

This change touches alot of code but its mostly
refactoring to remove factory service classes.

There is one behavior change: The model resource services
no longer require that the self link, and the "id" field match.

The client, in this case the resource allocation task, enforces
that.

Change-Id: I2ec3967272408cc35865661d61d6e9f7bf02fc8d