Adjust provision compute task and tests due to xenon changes
Xenon will no longer implicitly queue a request to a service that
does not exist. It does have a pragma however, that requests queuing
explicitly. Provision compute task and a test are now using this pragma.

ResourceAllocationTask maintains its high concurrency and continues
to create sub tasks in parallel, even before the compute services are
not yet created (all due to the use of queue pragma)

Change-Id: Ida244ef057c956f9a47858a28d31a443ff025e2c