Add a customProperties map to scheduled task
This will be used to 'tag' scheduled tasks so that they can
be managed. Note that we do not want to ue the documentSelfLink
as that is a global field used to ascertain uniqueness of the
service.

The propery map does not dictate uniqueness and multiple services
of the same type can be tagged with the same property

Change-Id: Id4ed99b32da8e0548457c4a2ee8b709a578d2ecc