Add support to just delete local resource representation
This option enables us to remove resources from the local
lucene store. This is useful when we have not created the
resources but are just tracking its state.

ResourceRemovalTaskService needs to support paginated queries.
That will be handled in a separate change

Support for making this change in the azure adapter will
also be handled separately

Change-Id: I0c9f8c8f8c1468d15bf2e116660dc4d580da0e94