Augment ScheduledTaskService to invoke task on service start
Before this change we create the scheduled task only when
the maintenance interval completes. This means that operations
like resource discovery won't result in resources being
dicovered for a significant amount of time. This change will
kick off the scheduled task once at startup and then honors
the maintenance interval

Change-Id: Ibbd9cb55f56b31aee775beeb1ec0492011b467ae