Do not mark photon model tasks as supporting idempotent POST
A task lifecycle should be updated via patches from the various
adapters. If a request for another instance of the task comes in
it should not be treated as a PUT and allowed to reset the
state of the current running task. Callers should see a failure
trying to POST and handle it.

This issue ws noticed during perf testing where periodic
enumeration tasks would kick off - if the previous one was not
complete, the service state would be reset due to the default
PUT handler resulting in the task state being clobbered

Change-Id: I0d8f7f20359d8b7b82776e082844d058416617d4